Revert "moving homeroom calculations to parent compnent"
This reverts commit f3b12bbb6d3b96601b4b2745abd1f1189937582b.
